requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
decorator==4.4.2
dnspython==2.1.0
et-xmlfile==1.1.0
idna==3.2
igramscraper==0.3.5
imageio==2.9.0
imageio-ffmpeg==0.4.4
instagram-scraper==1.10.2
joblib==1.0.1
moviepy==1.0.3
numpy==1.19.5
opencv-python==4.0.0.21
openpyxl==3.0.7
pandas==1.1.5
Pillow==8.3.1
proglog==0.1.9
PyAudio==0.2.11
pycparser==2.20
pymongo==3.12.0
python-dateutil==2.8.2
python-slugify==3.0.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
selenium==3.141.0
six==1.16.0
sklearn==0.0
sounddevice==0.4.2
text-unidecode==1.2
threadpoolctl==2.2.0
tqdm==4.62.1
urllib3==1.26.6
